Draft of a new runner api.
Kind of trouble, though.  The functional purity approach is... not the most helpful when you think of a distributed system future with networked APIs.

So... committing for amusement, but instantly going to turn around and start restructuring this to think more in terms of oplogs and resumables.  Golang makes it pleasant to pretend all operations are blocking, but that just doesn't survive networked concepts (or rather, it does, beautifully, but only when you build the blocking semantic behaviors over the resumable oplog parts).

Open question how to keep those separate, then.  I wanted the `def/act` package to be the API that "clients" would want, which does still mean the blocking bits.  Maybe all reader-type operations are blocking, and write/commands return the keys to start a reader-type, and that's the pattern to keep to.  I find it terribly tempting to combine them, but an API that allows both seems to become twice as complex.

Signed-off-by: Eric Myhre <hash@exultant.us>