Convert run command to go through new api.
Many improvements: Logging gets more consistent; the structural exit code reporting happens in a consistent place now; we've removed one more custom output byte funnel; and everything will be twice as good when the other cli commands just aim at this too.

The change to the error type in the legacy job result structure is to avoid the typed-nil-is-not-nil-nil problem.  Sigh.

Signed-off-by: Eric Myhre <hash@exultant.us>