Add benchmarks for stream mux.
It's expected that this system be slower than an in-memory buffer, of course, because it's fsyncing... but we should have an eye on it to make sure it's as good as reasonably possible within that constraint.

Just writing the first of these draws major attention to some close systems not available that just haven't bothered because we have no daemon modes in repeatr that give the descriptor leaks a chance to accumulate :/ todos in comments; may not bother fixing them until we have a use mode where they are in fact problems.  (This thing is gonna need nontrivial rewrites to work across processes anyway, frankly, which is likely coming up soon.)

Signed-off-by: Eric Myhre <hash@exultant.us>