Profile one specific test.
This library is rather neatish for this.  It's nice to be able to make the noise reduction up front -- it would be very difficult to do the same by filtering function trees explicitly because anything that starts goroutines could easily get misplaced that way; this captures and frames it fine.

Signed-off-by: Eric Myhre <hash@exultant.us>