Put a bufferer in the mux reader flow.
Debug printfs (how is it that these are always *so* productive, compared to the fancier things?) show that the codec library is pretty much dragging in one byte at a time far too often.  Buffering helps significantly.  (Hard to see this from the benchmarks of course, since we're feeding it events at a wall-clock-determined speed.)

Signed-off-by: Eric Myhre <hash@exultant.us>