Start implementing file transmat.  Single files.
There's no metadata available, and the files are actually... well, *files* as opposed to dirs.  This is a suprisingly big divergence -- bind mounts work different for files instead of dirs, almost all the places will have to change, etc.

This is useful for e.g. when there's a big dang ISO on the internet and you want it, and there's just no point in re-wrapping it as a tar.  That said, it's my feeling that you should use tars as a defacto standard on the internet at large, and metadata (hey there executable bit) is a good thing.

Leaving scan & upload with the single-file transmat as not yet implemented.  The level of abstraction shared with the other transmats is failing badly to DRY, and it should be cleaned up before it spreads any further.  ... And included: a big file o markdown on what should be coming up there.

Signed-off-by: Eric Myhre <hash@exultant.us>