Massage copyplacer until it works for files, as demonstrated by exercising the unpack command.
You can see how this is a bit lashy.

Also, removed O_EXCL from fs.PlaceFile.  I don't really think that's doing altogether that much good.  As a whole, that system is well into the "filesystems are not suitable for atomic thought" territory, and relies on systemically sane behavior for correctness; the exclusive create thing is a moderate irritant when we're now using it for single file operations which indeed do sometimes simply want to be an overwrite.

Signed-off-by: Eric Myhre <hash@exultant.us>