Make copyplacer DTRT even when the target dir doesn't exist yet.
This lets us remove a completely ridiculous crosscut we would otherwise need to create dirs when the transmat will want them, and files when the transmat does those.

There's probably a lot of places where this needs update for parity now, but the "right thing" from now on is to do one layer of creates.

Signed-off-by: Eric Myhre <hash@exultant.us>