Massive, total reimplementation of our git usage.
We no longer lean on the `git submodule` subcommands at all -- the entire process for each repo is handled by our own code.  This lets us cache working tree for each repo when there's a bunch of submodules, dramatical improving our ability to cache when the top level git commit changes, but shares many of its submodules -- by far the most common situation in practice! -- or, different repos share happen to use the same submodule commits, etc.

Submodule enumeration and handling is based on git's upstream scripts (the older(ish) ones that are bash and perl, because I can read them, frankly, but the C things on master seem to be not significantly different).  The parsing of the gitmodules file is based on the presumtion it's more or less INI format and ignoring the distinctions because they happen to be entirely places where we can get away with that.

The git transmat no longer expects to be used in tandem with a dircacher wrapping, because it has so many of its own internal layers to consider that the simplistic dircacher seems now like an unuseful diversion.

Git repo object dirs themselves are now also cached!  This is a best-effort sort of arrangement, bucketed based on remote URLs -- thus, doesn't successfully share objects if you're using a fallback list of warehouses -- but this is still a huge improvement in practical usage.

There's no parallelization for fetch, but we're close to it.  Similarly there's a lot of network traffic that could be elided by better cache checking before starting the expensive parts; will be stacking those enhancements on in future commits.

Signed-off-by: Eric Myhre <hash@exultant.us>