Migrate 'run' and 'twerk' subcommands to individual packages.  Refactor their shared parts.
Hitch now uses entirely meep errors (no more spacemonkey), and is correspondingly significantly better documented.  (Some of the errors raised by the 'def' package are not specifically aware of meep, but will fit in the same error handling defs anyway, which is very satisfying.)

Some serialization API behaviors specific to the run command are not yet migrated out of the core/cli package -- their single imported use is marked as "legacy"; we'll get to that in the next commit (this one is a big enough diff already!).

This is the last of the subcommands -- all are now migrated to individual packages!  Huzzah.

Signed-off-by: Eric Myhre <hash@exultant.us>