Rework entire run system to use the event stream all the time; it's not just for API mode anymore, it's a continuously used internal abstraction.
The upshot is mainly that API mode is now *really* trivial; it's just a loop polling the event API and spitting it out on the wire.  As it should be.

The `run` and `twerk` subcommands are now closer than before: both are consuming the same event stream, and using the same `terminal` package; just a tiny change in params to the terminal output setup.  Again, as it should be.  (Although there was a regression to twerk behavior which makes it always line buffer from a previous PR, and that has not yet been repaired.)

We're *probably* a good bit closer to being able to get rid of the current always-on output buffering behavior in executors, should we choose to do so.

Supervisors have started to appear in several places.  But the useful impact is still low; they haven't spread far enough to start making a coherent mesh of Betterness yet.

Oh, and the core/cli package we set out to refactor into oblivion yesterday is, in fact, now obliviated.  Phew.

Signed-off-by: Eric Myhre <hash@exultant.us>