Checkpoint during massive migration to API/serializable errors.
Will fail to compile.  About 4 of the 7 transmats are converted at this checkpoint; the other 3 cannot yet compile.

All major kinds of errors from rio are now exported from the `api/def` package.  This will make it possible not just to serialize them, but also build their proper deserialization into the API.  We'll have the ability for your consumer code to switch on error types by the end of this (provided you're using these same API packages in a golang client, of course).

The types should be roughly recognizable.  The names were standardized to the "Err{Something}" pattern, and a few of the overly-detailed types are being elided, but core concepts like "ErrHashMismatch" are essentially unchanged.

All these errors are being converted to meep.  Spacemonkey/errors imports are leaving the codebase at a rapid rate.  The meep use is conditional, though: we're still refraining from using meep directly in the `api/def` package.  Clients of those defs should not need to import or know anything about meep.  (If you go as far as `api/act`, then yes, meep will be there explicitly.)  But all the errors are real types, and thus `meep.TryPlan` with `ByType` routes is consistently a reasonable thing to use.

There are some fairly sweeping improvements to some of the transmats burried in the middle of this diff.  (Shhhhh.  I know, diffs should be one thing at a time.  But the refactors to errors have been occationally making it clear that one transmat has a vastly better flow than the other... so for example, the tar transmat is now further along the path of having the remote warehouse code abstracted, with ping functions, etc.)

Some errors are still not exported -- essentially, the group of internal errors from transmats that "shouldn't" happen on a correctly set up machine (basics! like, say, "not out of disk space").  The plan right now is to raise those as panics at the top level (e.g. when running `repeatr run` in "serialize"/API mode) rather than let them be retained on the RunRecord.Failure field and serialized there -- the reasoning being that we don't have the slightest desire to enshrine internal failure mode details in historical records or API types of any kind -- and that might need iteration, but we'll see how it goes.

Signed-off-by: Eric Myhre <hash@exultant.us>