Migrate placers to new error standards.
There are no API-type errors here; anything that goes wrong during a placer seems to be a "your host failed to support this build" scale of issue.  As with the assets package update, this thus isn't strictly critical path, but was affected by the refactor splash, so over to meep style it goes.

All tests under the `rio/...` packages pass again!  Yayy!

Signed-off-by: Eric Myhre <hash@exultant.us>