Move some critical bits of executor flow off of spacemonkey.
Not all of it; it's still raising spacemonkey errors because I'm trying to avoid broadening the scope any further.  This is just what was in the fairly direct line of fire to push compilability with the rio changes all the way up through here.

I abandoned the previously standing behavior for filtering to make sure errors raised past here were either rio subtypes, executor subtypes, or would be swaddled in an "unknown".  It's not clear that that forced translation layer ever did much good.  But moreover it's pretty hard to express without having a hierarchy available in quite the same way anymore.  Having a big whitelist is prospectively interesting, for the sake of self documenting code, but I can't think of a reasonably terse single-point-of-definition way to go about that, nor does the extra bits in stacks for the rest of time till kingdom come seem very reasonable.

Signed-off-by: Eric Myhre <hash@exultant.us>