Checkpoint while trying to add polymorphic marshal/unmarshal to errors attached to runrecords.
I am *very* unhappy with the capabilities of the serialization libraries I'm working with right now, and saving this checkpoint mostly to have it available as a reference in days to come when someone asks "why did you write another serialization library, instead of using %X?".  This.  This is why.  The right way to do these things is subtle; and this library is a swing and an ever-so-slight but ever-so-serious miss.

Signed-off-by: Eric Myhre <hash@exultant.us>