Rename `scan` subcommand to `pack`.
"Pack" more accurately describes what it's doing -- putting things into a packed-up warehouse-ready format that currently aren't, and getting the relevant hash type in the process -- and also is a more reasonable counterpart to "unpack" (nuff said).

Part of the trigger for this change is because we'd like to have a "scan" command that does verify and compute hash IDs for resources without necessarily packing or unpacking them.  As a result, the previous `repeatr scan` command is retired effectively immediately.  This is a breaking change; reminder that we're still alpha software with a 0.* version.

Signed-off-by: Eric Myhre <hash@exultant.us>