Raise additional errors from runc.
The fix appends another specific recognizer for this error, and a human-readable hint.  This keeps to the current standard of the surrounding code.  The desire to do a more aggressive refactor/upgrade/replacement of this area is strong, and important to do soon, but at least one lingering branch that was never able to be merged indicates it's more than an afternoon's work; maintaining a stable response to errors when the API beneath you refactors its own error codes is fraught with peril and needs a *lot* of regression testing.

This fixes #86.

Signed-off-by: Eric Myhre <hash@exultant.us>