Attempt to handle outputting a tar containing a single file.
Currently, such an output appears to work, and will contain the file content... but the tar emitted is incorrect: the mode bits say 0664, but the type is dir, and the name is "./" (the trailing slash is incorrect for a file).  More practically, GNU tar implementations will read the dir attributes, then skip the body that they don't believe should be there, and then end in error.

This solution is worse than the previous state though, and is only being committed for posterity/reference/comedy.  If the *paragraph* of doc warning the caller about the inconsistency in behavior for files wasn't enough of a warning sign: the immediate impact to our own code was some filepath.Join calls in the calling context getting *very* wonky, and correcting them would require continuing to propagate the special case for single files, which both seems wrong and results in proliferation of ugly extra params to quite a few methods.  There's got to be a better way.

Signed-off-by: Eric Myhre <hash@exultant.us>