Do a fresh sketch of what a re-imagined formula api would look like today, given insights we've gathered from working with the current format.
The highlights are almost entirely stuff already in the comments throughout the current def package: 'ware' is elevated to a type; warehouses are moved to a separate tree entirely; and so forth.

Also shown here is moving things like env vars to inputs.  This makes naming the inputs a bit more confusing, honestly, and so may not be a good choice.  The positive side is... env vars *are* an input.  (You could argue the Action is an input too, by this logic, though.)  This also reconciles awkwardly if you wanted to consider the Inputs section capable of emitting a filesystem tree and nothing else, which is a concept we've batted around many times (e.g. a whole separate tool, 'rio', that does repeatable IO, *without* containers, frequently seems like it would be useful).  So, yeah, this is... very debatable.

Wares are defined to have a simple string format.  This is how they would always be serialized.  Having a flat string instead of a pair of values is a huge simplicity consideration for the ecosystem -- for example, reppl has many internal map structures where the key is a ware ID, and these need to be serializable; thus reppl's life is vastly simpler if ware IDs are a simple type.

There's a fairly weighty comment block about what warehouse coord specification should look like.  This is a hard problem.  One of the things working with the current generation formulas has demonstrated over time is that updating warehouse coordinates is simply *annoying*.  As you produce larger pipelines, they're usually all *the same*, because they all just point to one big happy CAS bucket, and the copypasta is a useless distraction.  Or, they're for third-party URLs that *aren't* CAS, in which case they change every time the hash does, and that's ALSO a PITA to manage (reppl ended up choosing to store warehouse coords along with ware hashes to reduce this irritation).  So, we need something here that reduces this friction.  And it's also extremely clear that placing it directly in the Input struct is both an implementation pain (for hashing, etc) and also seems to require *more* explanation to new users (imagine painting the formula red or green based on whether it's part of the formula HID: why is it pinstripes?  awkward).  But knowing what doesn't work, while good (insert typical handwavey Edison quote here), does not make it clear what the excellent answer is: research, meditation and prototyping required.

Signed-off-by: Eric Myhre <hash@exultant.us>