Add 'Conjecture' struct to def2 outlines, to clarify where that is (and isn't).
Like the warehouse coordinate info: the experient with putting it inline with the output capture spec isn't *wrong* per se, but experience with that seems to indicate it might do better elsewhere.

Pros for separation:

- whether or not the reproduciblity conjecture includes an output slot definitely doesn't matter to the "curryed" hash, so it's good that it's not located with the rest of the output spec.

Cons:

- there is an argument to be made that given same entire setup, whether or not each output slot is expected to be reproducible really shouldn't be subject to a lot of debate.
- formula ends up with the output slot name repeated in another part of the doc.  mildly itchy.

Potentially more interesting is what impact the conjecture bits should have on the hash of the Results/RunRecord.  My feeling is that storing a RunRecord of *all* results, including the ones flagged as probably unreproducible (e.g. logs), is generally absurd to *not* do.  If you're checking reproducibility for $curriedHash in a table of [curriedHash->List:RunRecord]... yeah, yeah, ok, it's still pretty great if I could have a table that's bucketted by the hash of only the conjectured outputs.  So then: that table is defined by a conjecture configuration.  If those aren't included in the curriedHash, we need another join table; if they are included, the subject is done already.  I think that's the real tradeoff.

Signed-off-by: Eric Myhre <hash@exultant.us>