Readme and docs touches.
Build instruction update.  Lots of allusions to pipelining since we do indeedy have a solution for that now.  A variety of typo fixes.  Some TO-DOnes (priv dropping: a delightful thing that's been merged for quite some time; we ought to admit it!).  And state pretty boldly that repeatr is not a build tool, and does not compete with build tools: this is very, very (very! very!!) common question I get.  (Repeatr handles your $language-stdlib-n-compile version; build tools tend to punt to "somebody else".  We are stepping up to be that somebody else.)

Signed-off-by: Eric Myhre <hash@exultant.us>