Update recognizers for runc errors.
Using a table of string prefix and suffixes, since that's now apparently valid for all cases.

Most of the previous complaints about ambiguity remain.  Comments now discuss a little more specifically what it is we're aiming to get to.

Any error recognizer patterns that can't immediately be reproduced on my machine are being discarded: consider it mark-and-sweep GC of recognizer patterns.  This will almost certainly result in some regressions in error clarity.  This is not ideal, but with the lack of anything better than a stringly-typed API to runc, I don't know a better way to deal with this than mark-n-sweep and clean up again as opportunities present themselves.

Error output continues to be a problem: runc flushes its errors to stderr.  Immediately.  Which breaks such simple contracts as "stderr will be from the contained process".  I don't know what to do about this.  There are previous comments in this file about the same issue: apparently with the last version of runc, we fixed this by patching the runc sources.  That's a fairly frustrating thing to need... so I'm going to continue to hope I'm wrong and search for additional documentation on other ways to address this.

At a glance, we may be hosed: the first result for ddg'ing "runc stderr" is runc issue #349, opened by... me :( and still open.  https://github.com/opencontainers/runc/pull/1261 repeats the complaint, and offers a patch that appeared to have some traction, but is also still not merged.  The hunt continues...

Signed-off-by: Eric Myhre <hash@exultant.us>