Enable runc debug logging.
Forward it.

Was hoping that this would give us a clear indication of when a container has started successfully, so that we could use it for, say, flushing a buffer we've been accumulating for stderr.  No such luck; it appears to only emit news when a process ends, which is not very helpful.

AMEND: MAYBE!  Look at this: this is some sample logs from a "sleep 5" command:

```
INFO[04-11|14:33:24] Beginning execution!
DBUG[04-11|14:33:24] process exited                           runc-time=2017-04-11T14:33:24+02:00 runc-level=debug runc-pid=19070.000 runc-status=0.000
INFO[04-11|14:33:29] Execution done!                          elapsed=5.142
DBUG[04-11|14:33:29] process exited                           runc-status=0.000 runc-time=2017-04-11T14:33:29+02:00 runc-level=debug runc-pid=19071.000
```

I don't know if this is a detail I *want* to rely on, but apparently there's an auxilary process involved here which exits right away.  That *could* potentially be used as a trigger for opening the stderr buffer.

Signed-off-by: Eric Myhre <hash@exultant.us>