Shockingly, this is a solution to the runc stderr ambiguity problem.
A fragile, icky one, but functional so far as I can tell.  If we were supporting a matrix of runc versions, I would be 0% comfortable shipping this code, but since we are shipping a precisely known binary for runc, this might just be safe to pull off.

Let's still hope for better upstream fixes to runc issues 349, 1261, etc.

All compliance spec tests for the runc exectuor are green again.  I think we may be able to merge this one soon!

Signed-off-by: Eric Myhre <hash@exultant.us>