Remove Output.Conjecture flag.  Update documentation that used to refer to this concept.
This is being removed because it's not exactly the correct place for that concern, and worse, leaving the mentions there has proved very misleading.  Whether or not an output is consistent doesn't have much to do with how we run things.  Something should check that, sure; but how, and against what matrix you want to check it is a question that has more than one correct answer, and it *definitely* doesn't have anything to do with how we run something.

The Formula.SetupHash is a related concept, despite not actual covering results -- or perhaps because it specifically DOES NOT cover results, but DOES cover output slot specifications.  Most docs previously refering to the conjecture now make more sense referring to the SetupHash instead.

There is no function which currently computes a Formula.ResultsHash, which sounds like a function that would make more sense to concern itself with conjectures.  This doesn't exist largely because downstreams (namely, reppl, which is an example of a practical tool that's stitching things together and really cares about memoizing, etc) have so far been getting along fine by looking at each line of result hash separately.  In other words, there's no reason this function doens't exist; just that it hasn't been needed yet.  If we define one (and to be clear: it will not replace SetupHash; it will be ADifferentHash), then a conjecture-like field may make an appearance again, but until that time, removing it seems best.

Signed-off-by: Eric Myhre <hash@exultant.us>