Add additional comments documenting the semi-magical /dev behavior, and link to the relevant upstream source.
It does indeed look like runc would, at present, make it very difficult for us to opt out of this behavior.  We *could* populate our own /dev, and mount it in, but we would be required to do so *separately* from our existing filesystem assembly process, because the mount would need to be described to runc explicitly in their config rather than coasting in under the radar with the rest of our pre-assembled rootfs mount tree.

Signed-off-by: Eric Myhre <hash@exultant.us>