Change the printed version info
Removes build date because it messes with reproducibility.
Adds a tree hash which should be reproducible and is different
for different dirty builds.
Commit hash is now the full hash instead of the short hash.
Moves the dirty flag to not be attached to the commit hash.
Dirty is now triggered by untracked files.
Adds commit/author dates.

Signed-off-by: Calvin Behling <calvin.behling@gmail.com>