Drop some version info tools which make building more complicated.
Requiring git to build is... kind of irritating.  So we won't do it anymore, period.  The use of an env var to override it was a hack, irritating to notice when you forgot (because it would crash a build if you didn't have git on the path or were trying to build from a source export rather than full repo clone), and -- terminally -- it rather ruins the point since you're as likely to screw up passing that around as not.

Removing builddate for similar reasons.  See the reproducible-builds.org for a huge series of valid reasons why times and builds just don't go together if we want a better world.

Signed-off-by: Eric Myhre <hash@exultant.us>