clean reboot.
We've been stacking up notes for quite a while on what we'd like to change with APIs and schemas in what we've been apocryphally refering to as "v200".

This branch will be that v200.  Lots and lots of breaking changes and large refactors will be piled on.  Eventually, we'll merge it with the existing repeatr repo history, at which point that code base will be almost entirely replaced by this one.

Along the way we're going for modularization, much more through-and-through consistent use of *serializable* APIs at all points, more freedom to break up work into multiple *processes*, and improving both error handling and concurrency with advancements in libraries we've made since the first scratch at repeatr (there's been a lot of advancements to catch up on!).

This matches a similar work already going on in the hitch project.  That, in fact, is our trigger and our driver: we want a new `repeatr` command which can stand in for demos of that part of the toolchain.  At the same time, the extraction of I/O components to the `rio` library started some time ago, and we're going to honor that by starting top-down here: building the formula API, the executors, then a demo executor... gradually moving closer to the point where we'll need real 'rio' functions to do real work for us.  By the time we get there... we'll be done!

This may take a while.

Signed-off-by: Eric Myhre <hash@exultant.us>