api: meditations on where to separate setup/run.
Pretty sure by the end of that comment block we're on the right track but it took a while to be confident after throwing everything up in the air again like this.

Also one of the major trickybits is figuring out where we want to put the warehouseaddr/discovery info in the v200 API, since we just know we *don't* want it too central in formulas anymore.  So, cascade of meditations on that coming up next.....

Signed-off-by: Eric Myhre <hash@exultant.us>