api: meditations on warehouse/discovery config.
... a LOT of meditations.

We've had this on the back burner for quite a while.  We know we need to do SOMETHING about it for the v200 change -- one of the biggest issues we have with the current formulas is, by far, the silo URL fields and where we put them, for half a dozen reasons -- but what, as always, is a hard question.

This is pretty much stream-of-consciousness, so it's a large wall of text.  I think we got somewhere by the end.  I'll drop most of it again shortly.

Next: try to draft all that thoughtwank into actual structs, eh?

I think there's actually going to be more than one of them...

Signed-off-by: Eric Myhre <hash@exultant.us>