api/repeatr: RunFunc warehouse params: addr only.
All the warehouse options go via the function params; comments removed in this diff about them sneaking along in the formula pointer were just Wrong.  (It may well be that we tolerate them residing in the same *file* when serialized sometimes, but the formula object in v200 shall consistently have *no fields* which are not covered by "setupHash": that is a victory criteria.)

We're not going to pass the full warehouse configuration here.  I waffle on this, and review may be needed, and iteration almost certainly so.  The current reasoning I'm operating by is: supposing you were using this API to invoke repeatr runs in a remote k8s cluster: it is not necessarily desirable to pass all the necessary warehouse credentials over the wire here; more likely, you've provided those secrets to your remote run environment in some other secured and auditable channel, and we should only be telling you *which ones* to use (perhaps by e.g. "user@" patterns in the URL, if that is of consequence).

Signed-off-by: Eric Myhre <hash@exultant.us>