api: package alias.
Mirrors a relocation already done in the rio source tree.

We'll almost certainly end up publishing this package in its own source tree, later.  Right now, we're going to keep treating the one in this, the r200 repo, as canonical, and develop it here for a while; but once development rolls along enough for the package to stabilize a bit, we'll almost certainly be pulling it out into its own submodule.

Signed-off-by: Eric Myhre <hash@exultant.us>