executor: first draft of demo executor.
Returns a fairly trivial runrecord.

We're choosing to include the real hostname -- yes.  This struct also includes things like time, and we haven't seen a fit reason to fake or fixturize that either -- the intention is to use this executor in large integration tests, not unit tests, so we actually *want* it to express the full range of behaviors, like always-novel runrecord emission.

Add our first error category of the repeatr API package.

Add ExitCode to runrecord.  The current repeatr behavior overloads the output map with a magic-keyed entry that contains a "hash" that's really an int.  This is turned out to feel fairly ridiculous, and we're going to give it a clearer space this time around.  (I think the original idea there was to have nothing *but* that map be retained in most records, but that seems... to not have happened like that anyway, so let's give up that ghost?)

Signed-off-by: Eric Myhre <hash@exultant.us>