Switch to the public extracted api repo.
Most of the diff is due to the slight package rename (our placeholder earlier did not take into account having a "go-" prefix on the repo name, a convention we're sort of trying on for size here.)

This should leave us with approximately zero blockers to starting to gracefully use rio in external processes from here.  Wahoo!

Signed-off-by: Eric Myhre <hash@exultant.us>