link rio as lib.
The plan is for Repeatr to use some parts of the Rio source base *as a library*, even while also running some operations -- namely, all packs and unpacks -- as fork&execs of a separate rio executable process.

This will give us the best of both worlds: the I/O heavy work is done in subprocesses (which can be terminated reliably at the OS signally level, as well as audited, nice'd, etc separately by a sysadmin); and at the same time, control logic which would be both a big PITA to make entirely RPC'able and also quite useless to spawn another intermediary process for: that we just run in-process with no overhead or fuss.

Specifically, we're going to use the whole stitch/assembler components as funcs in repeatr, because it's useful and sensible to do so (even though the code resides in the rio repository since the rio commands also have unpack-tree functionality when used standalone)... while the UnpackFunc used to do the heavy lifting will be a `rioexecclient.UnpackFunc`, spawning new processes to do the heavy work.

Builds of the Repeatr binary will end up with some packages compiled in which are redundant with the ones compiled into a Rio binary -- namely, the 'stitch' package -- but it's *far* from total duplication: notice that the 'stitch' package does *NOT* have any transitive dependencies on any of the transmat code, warehouse code, etc.  (Even the cache features are split into two separate packages in Rio: one for behavioral description, and one for actual implementation.  This is why; Repeatr will end up linking in *only the former*.)

Signed-off-by: Eric Myhre <hash@exultant.us>