executor: sketch chroot impl outline.
I'd like to try much harder to reduce the amount of prefix and suffix copypasta in executor implementations this time... a lot of the IO work might all end up in a helper function that takes a `func(FormulaActions, InputControl, Monitor) error`, or something.

It's nice to realize that the whole assembler unit already nicely ties off all the unpackTool and caching configuration stuff into one nice self-contained bubble.

Signed-off-by: Eric Myhre <hash@exultant.us>