rio: now builds within repeatr's build; needed.
There are a number of ways to go about this, and all of them suck in various ways and degrees:

- Option 1: ignore the problem, and rely on host $PATH.  Total non-starter;
  we have tests in raceway which will instantly flag this (as they should).

- Option 2: use a timeless stack pipeline to build rio in another container,
  then pipe the ware into the repeatr build step.  This is AWESOME,
  but would be a shit bootstrapping cycle, so we're not going to.
  (The Timeless repo omnibus builds will probably do this, though.)

- Option 3: `(cd .gopath/src/go.polydawn.net/rio && goad install)`.
  This would work, and moves all our bootstrap issues into the realm of
  git, thus breaking the cycle and being a valid choice.  However, this
  implementation implies recursive submodules, and those suck fiercely.

- Option 4 -- what we're actually doing: build Rio *in* Repeatr's
  source tree.  This involves essentially ignoring Rio's build script
  (but that's fine in practice because it's so trivial), and also means
  we have to add all libraries Rio depends on as submodules in Repeatr
  (but this is also a fairly small diff in practice).

So, Option 4 it is.

This also plays surprisingly nicely with the rest of our toolchain: if you're, say, running both the Rio and Repeatr tests in Raceway, they'll be sharing caches of all the library git repos, etc.

Later, we might need to conditionalize this Rio build, etc, since again, the Timeless Stack omnibus builds and complete suite releases will probably use a different build lifecycle, and this build will become redundant.  But this is a series of optimization issues we can kick down the road.

Signed-off-by: Eric Myhre <hash@exultant.us>