fixture: save busybox+bash blob.
The WareID is "tar:6q7G4hWr283FpTa5Lf8heVqw9t97b5VoMU6AGszuBYAz9EzQdeHVFAou7c4W9vFcQ6".

This is a binary, and git should not contain a ton of these, but we're going to simply make this a fixture that we stick with for a LOOOONG time.  Though there's not much here, we can use a series of bash hacks to cheaply simulate all kinds of fun things inside a container (e.g. use a bunch of echo-redirects and mkdirs to make more fixtures, etc etc etc).

The blob was originally produced by asking the Alpine tools to give us bash and busybox, which results in a rather compact but nice and usable image.  It was hash "l4VD4VM5geqNM_QY8eRA1gPUYiZHQJrXjyYYsp7W5IJUuOOjKytwe2TWSohb4QYs" in the *old* scheme, and originally produced by formulas in the formulary.git repo, and used for rix bootstrap experiments last spring (so, it's already a tad out of date, doubtless, but we're not exactly running untrusted network input through it in our tests, so I can't imagine an earthly reason to care).  I'd be a little more thorough about strongly linking and hashing the provenence if it helped; however, it does not; this image already does not rebuild reproducibly, because the network state that the Alpine package manager consumes has (already, some time ago) changed.  If we want reproducible images here, we need to make them... ourselves.  Essentially all other tools have failed.

Executor tests will begin using this momentarily.

Signed-off-by: Eric Myhre <hash@exultant.us>