executor: assertions on exit codes and round-trip rootfs hash (!!!)
It's really exciting that this worked pretty much first shot.  There was a loooooot of iteration and chasing of mtimes and general discovery of pain-in-the-patootie when we were at this stage in implementing the first generation of Repeatr.  We actually learned things!  Second system syndrome is... going pretty well; not so much a syndrome.

Signed-off-by: Eric Myhre <hash@exultant.us>