cli: more args parsing, bindings, and test setup.
I'm going for some fairly esoteric code style here, but the aims are:

- get the struct defs to appear exactly once,
- get the defs to *not go very far*,
- keep the args binding setup calls *very* close to the relevant structs,
- keep the func that turns the struct into a Real Stuff call close to both of them,
- and be very, very testable.

The first four are pretty hard given this library's API, to be honest.  But we seem to be doing pretty well here.

The "be testable" part is also nontrival.  I'm trying something new where we return the parsed args before running anything, so we can write tests that inspect it.

This object for test inspection is a little wild because A) it has no reason to exist in the real main code paths and B) its type safety is 0% but C) that *doesn't matter* because the main code *doesn't actually use it*; it's just hanging things on the struct so that they just *happen* to be visible to the test harness when it's the caller.

The test code doesn't really end up with much of a way to inspect the structs with types, either, which is odd.  But I think with the level of vision we want here, chugging the things back through a json serializer and doing big stinky string diffs should be... fine, honestly.  (The alternative is that the struct definitions have to move up to top-level declarations, and that makes the other objectives about keeping the binding methods closeby much harder.)

Signed-off-by: Eric Myhre <hash@exultant.us>